+++ structurizr-core/src/com/structurizr/view/Styles.java
@@ -61,15 +61,6 @@ public final class Styles {
         this.relationships = new LinkedList<>();
     }
 
-    public void addDefaultStyles() {
-        addElementStyle(Tags.ELEMENT).shape(Shape.RoundedBox);
-        addElementStyle(Tags.SOFTWARE_SYSTEM).background("#1168bd").color("#ffffff");
-        addElementStyle(Tags.CONTAINER).background("#438dd5").color("#ffffff");
-        addElementStyle(Tags.COMPONENT).background("#85bbf0").color("#000000");
-        addElementStyle(Tags.PERSON).background("#08427b").color("#ffffff").shape(Shape.Person);
-        addElementStyle(Tags.INFRASTRUCTURE_NODE).background("#ffffff");
-    }
-
     public Collection<RelationshipStyle> getRelationships() {
         return relationships;
     }